Don't panic on dropped frames
On some lower end systems it's possible in some high pressure moments
hit an edge case where a frame is dropped whilst buffers are still
waiting to be mapped, causing wgpu to signal an error on the mapping
callback and the profiler to panic.

This is fixed by tracking wether the current frame was dropped or not
and ignoring buffer mappings if it was dropped.